+++ src/main.rs
@@ -10,12 +10,12 @@ fn setup_database() -> Result<(), Box<dyn std::error::Error>> {
         "CREATE TABLE IF NOT EXISTS wdi (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             country TEXT,
-            fertility_rate INTEGER,
-            viral INTEGER,
-            battle INTEGER,
-            cpia_1 INTEGER,
-            cpia_2 INTEGER,
-            debt INTEGER
+            fertility_rate REAL,
+            viral REAL,
+            battle REAL,
+            cpia_1 REAL,
+            cpia_2 REAL,
+            debt REAL
         )",
         [],
     )?;
